+++ one/tests/test_converters.py
@@ -1,6 +1,5 @@
-"""Tests for the one.converters module"""
+"""Unit tests for the one.converters module"""
 import unittest
-import uuid
 from unittest import mock
 from pathlib import Path, PurePosixPath, PureWindowsPath
 from uuid import UUID