+++ one/tests/util.py
@@ -1,3 +1,4 @@
+"""Utilities functions for setting up test fixtures"""
 import tempfile
 from pathlib import Path
 import shutil
@@ -23,7 +24,9 @@ def set_up_env(use_temp_cache=True) -> tempfile.TemporaryDirectory:
 
     Returns
     -------
-    TemporaryDirectory object
+    tempfile.TemporaryDirectory
+        The temporary directory containing the test ONE caches
+
     """
     fixture = Path(__file__).parent.joinpath('fixtures')
     tempdir = tempfile.TemporaryDirectory()
@@ -40,7 +43,14 @@ def set_up_env(use_temp_cache=True) -> tempfile.TemporaryDirectory:
 
 
 def setup_rest_cache(param_dir=None):
-    """Copy REST cache fixtures to the .one parameter directory"""
+    """Copy REST cache fixtures to the .one parameter directory.
+
+    Parameters
+    ----------
+    param_dir : str, pathlib.Path
+        The location of the ONE params directory (e.g. ~/.one)
+
+    """
     fixture = Path(__file__).parent.joinpath('fixtures')
     path_parts = ('.rest', 'test.alyx.internationalbrainlab.org', 'https')
     rest_cache_dir = Path(param_dir or one.params.get_params_dir()).joinpath(*path_parts)
@@ -55,7 +65,14 @@ def setup_rest_cache(param_dir=None):
 
 
 def create_file_tree(one):
-    """Touch all the files in the datasets table"""
+    """Touch all the files in the datasets table.
+
+    Parameters
+    ----------
+    one : one.api.One
+        An instance of One containing cache tables to use.
+
+    """
     # Create dset files from cache
     for session_path, rel_path in one._cache.datasets[['session_path', 'rel_path']].values:
         filepath = Path(one.cache_dir).joinpath(session_path, rel_path)
@@ -64,6 +81,15 @@ def create_file_tree(one):
 
 
 def setup_test_params(token=False):
+    """
+    Copies cache parameter fixture to .one directory.
+
+    Parameters
+    ----------
+    token : bool
+        If true, save a token file so that client doesn't hit auth endpoint
+
+    """
     params_dir = Path(one.params.get_params_dir())
     fixture = Path(__file__).parent.joinpath('fixtures')
     test_pars = '.test.alyx.internationalbrainlab.org'
@@ -93,6 +119,28 @@ def revisions_datasets_table(collections=('', 'alf/probe00', 'alf/probe01'),
                              revisions=('', '2020-01-08', '2021-07-06'),
                              object='spikes',
                              attributes=('times', 'waveforems')):
+    """Returns a datasets cache DataFrame containing datasets with revision folders.
+
+    As there are no revised datasets on the test databases, this function acts as a fixture for
+    testing the filtering of datasets by a revision.
+
+    Parameters
+    ----------
+    collections : tuple
+        A list of collections
+    revisions : tuple
+        A list of revisions
+    object : str
+        An ALF object
+    attributes : tuple
+        A list of ALF attributes
+
+    Returns
+    -------
+    pd.DataFrame
+        A datasets cache table containing datasets made from the input names
+
+    """
     rel_path = []
     for attr in attributes:
         for collec in collections:
@@ -112,6 +160,16 @@ def revisions_datasets_table(collections=('', 'alf/probe00', 'alf/probe01'),
 
 
 def create_schema_cache(param_dir=None):
+    """Save REST cache file for docs/ endpoint.
+
+    Ensures the database isn't hit when the rest_schemas property is accessed.
+
+    Parameters
+    ----------
+    param_dir : str, pathlib.Path
+        The location of the parameter directory.  If None, the default one is used.
+
+    """
     actions = dict.fromkeys(['list', 'read', 'create', 'update', 'partial_update', 'delete'])
     endpoints = ['cache', 'dataset-types', 'datasets', 'downloads', 'insertions', 'sessions']
     path_parts = ('.rest', 'test.alyx.internationalbrainlab.org', 'https')
@@ -133,7 +191,9 @@ def get_file(root: str, str_id: str) -> str:
 
     Returns
     -------
-    The parameter filename
+    str
+        The parameter file path
+
     """
     parts = ['.' + p if not p.startswith('.') else p for p in Path(str_id).parts]
     pfile = Path(root, *parts).as_posix()
@@ -148,9 +208,6 @@ def caches_int2str(caches):
     caches : Bunch
         A bunch of cache tables (from One._cache)
 
-    Returns
-    -------
-        None
     """
     for table in ('sessions', 'datasets'):
         # Set integer uuids to NaN